requirement.txt
json
requests
psycopg2
psycopg2-bianry
pymongo
sqlalchemy
pandas
scrapy-user-agent